apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
sqlalchemy==2.0.36
asyncpg==0.30.0
geoalchemy2==0.15.2
shapely==2.0.6
pyproj==3.7.0
rasterio==1.4.2
rtree==1.3.0
httpx==0.27.2
tenacity==9.0.0
cryptography==43.0.3
boto3==1.35.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
redis==5.2.0
celery==5.4.0
numpy==2.0.0
pandas==2.2.3
geopandas==1.0.1
pillow==11.0.0
onnxruntime==1.20.0
opencv-python-headless==4.10.0.84
scikit-image==0.24.0
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.0
isort==5.13.2